tests/test_pdf_to_json.py
import pytest
from pathlib import Path
from app.core.pdf_to_json import PdfToJsonConverter

class TestPdfToJsonConverter:
    """
    Comprehensive test suite for PDF to JSON conversion
    """
    @pytest.fixture
    def sample_pdf_path(self, tmp_path):
        """
        Create a sample PDF for testing
        """
        sample_pdf = tmp_path / "sample.pdf"
        # Implement PDF creation logic
        return sample_pdf

    def test_successful_conversion(self, sample_pdf_path, tmp_path):
        """
        Test successful PDF to JSON conversion
        """
        output_path = tmp_path / "test_output.json"
        result = PdfToJsonConverter.convert(
            str(sample_pdf_path), 
            str(output_path)
        )
        
        assert result is True
        assert output_path.exists()
        assert output_path.stat().st_size > 0

    def test_invalid_pdf_conversion(self, tmp_path):
        """
        Test conversion with invalid PDF
        """
        output_path = tmp_path / "invalid_output.json"
        result = PdfToJsonConverter.convert(
            "/path/to/non/existent/file.pdf", 
            str(output_path)
        )
        
        assert result is False